Add Dockerfile
This puts the web application in a Docker container. It runs independently. The next step is to create a docker-compose file, and add in the database, so we can start developing on a grown-up database like Postgres rather than SQLite.
To run this container, use the following command:
docker build -t <image_tag> . && docker run -p 5000:5000 <image_tag>
-p 5000:5000
<image_tag>