Update migration with missing foreign key
When flask-migrate tried to drop the column 'masterclass_id', it first had to drop the foreign key constraint. This wasn't autogenerated, and so when I tried to run the upgrades it crapped out. I've fixed this migration up with the right name and you should now be able to run migrations without issue.