Rewrite 'index' link in base template
I noticed that when trying to click the 'Masterclasses' link in the header from the view at masterclass/<masterclass_id>, there was an error. I've fixed this error by changing this link to 'url_for', which means Flask will look for the route labelled 'index' rather than just appending 'index' to the current path.