Use the app factory pattern
This is a significant change. To enable testing, I've refactored the application to use an 'app factory' pattern. What this means is that the application is now created by calling the 'create_app' method in app/__init__. This in turn allows us to create the app with different configurations for different scenarios - in the simplest case, for running in a development mode versus a test mode.

Using this pattern opens the door to using blueprints, which allow us to group routes together in logically similar patterns. Advanced use of blueprints allows us to add paths to every route in the blueprint and thereby separate them - for example, allowing for '/user/create' and '/masterclass/create'. Both are served by a route called "create", but because they're under different blueprints, there's no clash. I can walk you through an example when I see you next.