Set up test fixtures
First I've added pytest, to allow for testing. I've then added three fixtures and the Configs necessary to make them work.
TestConfig turns off 'login_required'. This is really helpful for unit testing, as it means we don't need to hack a workaround just to test that our routes are  working as they should. This TestConfig is used when setting up the 'client' fixture, which is the object we use when testing whether things are working. You can see an example in 'test_login_not_required'. Since there's no login, I've also had to set up a database - the client goes straight to the index. The index route makes a call to the database, so in order for the test to pass there first has to be a database.

The other test case uses the 'login_required_client', which is a test client configured with logins on. This is really helpful when testing that we're sure the areas that should be behind logins are!